requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
frozenlist==1.5.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
multidict==6.1.0
numpy==2.1.2
propcache==0.2.0
py-aiowialon==1.3.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-wialon==1.2.4
simplejson==3.19.3
typing_extensions==4.12.2
yarl==1.17.0
